+++ projects/core/src/main/kotlin/site/siredvin/peripheralium/api/blockentities/IOwnedBlockEntity.kt
@@ -1,0 +1,7 @@
+package site.siredvin.peripheralium.api.blockentities
+
+import net.minecraft.world.entity.player.Player
+
+interface IOwnedBlockEntity {
+    var player: Player?
+}