+++ projects/core/src/main/kotlin/site/siredvin/peripheralium/common/blocks/BaseTileEntityBlock.kt
@@ -3,6 +3,9 @@ package site.siredvin.peripheralium.common.blocks
 import net.minecraft.core.BlockPos
 import net.minecraft.server.level.ServerLevel
 import net.minecraft.util.RandomSource
+import net.minecraft.world.entity.LivingEntity
+import net.minecraft.world.entity.player.Player
+import net.minecraft.world.item.ItemStack
 import net.minecraft.world.level.Level
 import net.minecraft.world.level.block.BaseEntityBlock
 import net.minecraft.world.level.block.Block
@@ -14,6 +17,7 @@ import net.minecraft.world.level.block.entity.BlockEntityType
 import net.minecraft.world.level.block.state.BlockState
 import net.minecraft.world.level.material.Material
 import site.siredvin.peripheralium.api.blockentities.IObservingBlockEntity
+import site.siredvin.peripheralium.api.blockentities.IOwnedBlockEntity
 import site.siredvin.peripheralium.api.peripheral.IPeripheralTileEntity
 
 abstract class BaseTileEntityBlock<T: BlockEntity>(
@@ -73,6 +77,15 @@ abstract class BaseTileEntityBlock<T: BlockEntity>(
             if (tile is IObservingBlockEntity)
                 tile.placed()
         }
+    }
+
+    override fun setPlacedBy(level: Level, pos: BlockPos, state: BlockState, entity: LivingEntity?, stack: ItemStack) {
+        super.setPlacedBy(level, pos, state, entity, stack)
+        val tile = level.getBlockEntity(pos)
+        if (tile is IOwnedBlockEntity && !level.isClientSide && entity is Player)
+            tile.player = entity
+        if (tile is IObservingBlockEntity)
+            tile.setPlacedBy(entity, stack)
     }
 
     override fun onRemove(