+++ projects/core/src/main/kotlin/site/siredvin/peripheralium/computercraft/peripheral/owner/BlockEntityPeripheralOwner.kt
@@ -8,13 +8,17 @@ import net.minecraft.world.entity.player.Player
 import net.minecraft.world.item.ItemStack
 import net.minecraft.world.level.Level
 import net.minecraft.world.level.block.entity.BlockEntity
-import site.siredvin.peripheralium.common.blocks.GenericBlockEntityBlock
+import net.minecraft.world.level.block.state.properties.DirectionProperty
+import site.siredvin.peripheralium.api.blockentities.IOwnedBlockEntity
 import site.siredvin.peripheralium.api.peripheral.IPeripheralTileEntity
+import site.siredvin.peripheralium.api.storage.ExtractorProxy
 import site.siredvin.peripheralium.api.storage.SlottedStorage
+import site.siredvin.peripheralium.common.blocks.GenericBlockEntityBlock
 import site.siredvin.peripheralium.util.DataStorageUtil
+import site.siredvin.peripheralium.util.world.FakePlayerProviderBlockEntity
 import java.util.*
 
-class BlockEntityPeripheralOwner<T>(val tileEntity: T) :
+class BlockEntityPeripheralOwner<T>(private val tileEntity: T, private val facingProperty: DirectionProperty = GenericBlockEntityBlock.FACING) :
     BasePeripheralOwner() where T : BlockEntity, T : IPeripheralTileEntity {
 
     override val level: Level?
@@ -26,29 +30,34 @@ class BlockEntityPeripheralOwner<T>(val tileEntity: T) :
         get() = tileEntity
 
     override val facing: Direction
-        get() = tileEntity.blockState.getValue(GenericBlockEntityBlock.FACING);
+        get() = tileEntity.blockState.getValue(facingProperty);
 
     override val owner: Player?
-        get() = null
+        get() = (tileEntity as? IOwnedBlockEntity)?.player
     override val dataStorage: CompoundTag
         get() = DataStorageUtil.getDataStorage(tileEntity)
 
-    override val storage: SlottedStorage?
-        get() = null
+    override val storage: SlottedStorage? by lazy {
+        ExtractorProxy.extractStorage(tileEntity.level!!, tileEntity.blockPos) as? SlottedStorage
+    }
 
     override fun markDataStorageDirty() {
         tileEntity.setChanged()
     }
 
     override fun <T> withPlayer(function: (ServerPlayer) -> T, overwrittenDirection: Direction?, skipInventory: Boolean): T {
-        throw RuntimeException("Not implemented yet")
+        if (tileEntity !is IOwnedBlockEntity)
+            throw IllegalArgumentException("Cannot perform player logic without owned block entity")
+        return FakePlayerProviderBlockEntity.withPlayer(tileEntity, function, overwrittenDirection = overwrittenDirection, skipInventory = skipInventory)
     }
 
     override val toolInMainHand: ItemStack
         get() = ItemStack.EMPTY
 
     override fun storeItem(stored: ItemStack): ItemStack {
-        throw RuntimeException("Not implemented yet")
+        if (storage == null)
+            return stored
+        return storage!!.storeItem(stored)
     }
 
     override fun destroyUpgrade() {