+++ projects/forge/src/main/kotlin/site/siredvin/peripheralium/forge/ForgePeripheraliumPlatform.kt
@@ -13,6 +13,7 @@ import net.minecraft.core.Holder
 import net.minecraft.core.HolderSet
 import net.minecraft.core.Registry
 import net.minecraft.nbt.CompoundTag
+import net.minecraft.nbt.Tag
 import net.minecraft.resources.ResourceKey
 import net.minecraft.resources.ResourceLocation
 import net.minecraft.server.level.ServerLevel
@@ -177,5 +178,9 @@ class ForgePeripheraliumPlatform: PeripheraliumPlatform {
 
     override fun getPocketUpgrade(key: String): IPocketUpgrade? {
         return PocketUpgrades.instance().get(key)
+    }
+
+    override fun nbtToLua(tag: Tag): Any? {
+        return NBTUtil.toLua(tag)
     }
 }